requirements.txt
# 출결 관리 자동 캡처 프로그램 - 필수 라이브러리
# Python 3.10.11

# 화면 캡처
mss==9.0.1

# 이미지 처리
Pillow==10.1.0
numpy==1.24.3

# 얼굴 감지 (InsightFace)
insightface==0.7.3
onnxruntime==1.16.3

# GUI (tkinter는 Python 내장)
# 스케줄링 (자체 구현)
# CSV (Python 내장)

# 테스트 (개발 시에만 필요)
pytest==7.4.3
